Fuse dual updates if operations can be saved
If we are newly threatnening a piece that gets removed by the next move, or get threatened by the piece that gets removed next turn,
we can omit the addition of that threat and the removal of it to save one add and one sub.

In theory this observation should also work for threats from/to the square the next piece moves from, but I did not get that to work yet.

Thanks to @cj5716 for pointing out that addDirtyThreat can be templated

Passed STC:
LLR: 2.95 (-2.94,2.94) <0.00,2.00>
Total: 770464 W: 198015 L: 196671 D: 375778
Ptnml(0-2): 2336, 83268, 212756, 84460, 2412
https://tests.stockfishchess.org/tests/view/68f4e178637acd2a11e72170

bench 2702035